🛠결과_간단 모듈화
