+++ pages/result/components/AgeBarChart.tsx
@@ -1,25 +1,11 @@
-import { SurveyData } from "@/pages/edit-survey/editInterface";
 import Chart from "react-apexcharts";
-import { SurveyResult } from "../[id]";
-
-interface BarChartProps {
-  item: SurveyData; // 설문 데이터
-  surveyResult: { [key: string]: SurveyResult }; // 설문 결과 데이터
-}
+import { ChartProps } from "./resultInterface";
+import { engAgeCategories } from "./categories";
 
 interface ResponseProps {
   questionId: number; // 질문 ID
   request: { answers: string | string[] }; // 응답 내용
 }
-
-export const ageCategories = [
-  "UNDER_TWENTY",
-  "TWENTIES",
-  "THIRTIES",
-  "FOURTIES",
-  "FIFTIES",
-  "OVER_SIXTIES",
-];
 
 export const engToKorAge: { [key: string]: string } = {
   UNDER_TWENTY: "20대 미만",
@@ -30,12 +16,12 @@ export const engToKorAge: { [key: string]: string } = {
   OVER_SIXTIES: "60대 이상",
 };
 
-export const AgeBarChart = ({ item, surveyResult }: BarChartProps) => {
+export const AgeBarChart = ({ item, surveyResult }: ChartProps) => {
   // 각 설문 옵션에 대한 응답 수를 계산하여 series 데이터 생성
   const series = item.options.map((option) => ({
     name: option.answer, // 각 옵션의 답변 내용
     // 나이별로 응답 수를 계산
-    data: ageCategories.map((age) => {
+    data: engAgeCategories.map((age) => {
       // result 객체는 각 응답자의 응답을 담고 있음
       return Object.values(surveyResult).reduce((total, result) => {
         // 응답자의 나이가 현재 카테고리와 일치할 경우
@@ -70,7 +56,7 @@ export const AgeBarChart = ({ item, surveyResult }: BarChartProps) => {
     },
     xaxis: {
       // x축 카테고리 설정
-      categories: ageCategories.map((age) => engToKorAge[age]),
+      categories: engAgeCategories.map((age) => engToKorAge[age]),
     },
     legend: {
       position: "bottom" as const, // 범례
