+++ pages/result/components/categories.ts
@@ -1,0 +1,10 @@
+export const engAgeCategories = [
+  "UNDER_TWENTY",
+  "TWENTIES",
+  "THIRTIES",
+  "FOURTIES",
+  "FIFTIES",
+  "OVER_SIXTIES",
+];
+
+export const engGenderCategories = ["MALE", "FEMALE", "UNKNOWN"];
